requirements.txt
werkzeug
flask
gunicorn
pillow
pyautogui